Refactor downloader module from single file into package
Split the monolithic downloader.py (~1170 lines) into a package with
logical sub-modules by media type: core.py (utilities/constants),
audio.py, images.py, video.py, text.py, and documents.py.
All symbols re-exported from __init__.py for backward compatibility.
