+++ vtsearch/datasets/downloader/images.py
@@ -1,0 +1,325 @@
+"""Image dataset downloaders: CIFAR-10, Caltech-101/256, Oxford Flowers, Food-101, EuroSAT, Stanford Dogs."""
+
+import tarfile
+import zipfile
+from pathlib import Path
+from typing import Optional
+
+from vtsearch.config import DATA_DIR
+from vtsearch.datasets.downloader.core import (
+    CALTECH101_DOWNLOAD_SIZE_MB,
+    CALTECH101_URL,
+    CALTECH256_DOWNLOAD_SIZE_MB,
+    CALTECH256_URL,
+    CIFAR10_DOWNLOAD_SIZE_MB,
+    CIFAR10_URL,
+    EUROSAT_DOWNLOAD_SIZE_MB,
+    EUROSAT_URL,
+    FOOD101_DOWNLOAD_SIZE_MB,
+    FOOD101_URL,
+    IMAGE_DIR,
+    OXFORD_FLOWERS_DOWNLOAD_SIZE_MB,
+    OXFORD_FLOWERS_LABELS_URL,
+    OXFORD_FLOWERS_URL,
+    STANFORD_DOGS_DOWNLOAD_SIZE_MB,
+    STANFORD_DOGS_URL,
+    ProgressCallback,
+    _default_progress,
+    _download_and_extract,
+    download_file_with_progress,
+)
+
+
+def download_cifar10(on_progress: Optional[ProgressCallback] = None) -> Path:
+    """Download and extract the CIFAR-10 image classification dataset.
+
+    Downloads ``cifar-10-python.tar.gz`` from the configured ``CIFAR10_URL``
+    into ``DATA_DIR`` if it is not already present, then extracts it and deletes
+    the archive to reclaim disk space. Both steps report progress via
+    *on_progress*.
+
+    Args:
+        on_progress: Optional progress callback. Falls back to the
+            application-wide ``update_progress`` when ``None``.
+
+    Returns:
+        Path to the ``cifar-10-batches-py/`` directory containing the raw pickle
+        batch files (e.g. ``data/cifar-10-batches-py``).
+    """
+    if on_progress is None:
+        on_progress = _default_progress()
+
+    extract_dir = DATA_DIR / "cifar-10-batches-py"
+    _download_and_extract(
+        url=CIFAR10_URL,
+        archive_name="cifar-10-python.tar.gz",
+        extract_to=DATA_DIR,
+        check_path=extract_dir,
+        download_size_mb=CIFAR10_DOWNLOAD_SIZE_MB,
+        dataset_name="CIFAR-10",
+        on_progress=on_progress,
+    )
+    return extract_dir
+
+
+def download_caltech101(on_progress: Optional[ProgressCallback] = None) -> Path:
+    """Download and extract the Caltech-101 image classification dataset.
+
+    Downloads ``caltech-101.zip`` from the configured ``CALTECH101_URL``
+    into ``DATA_DIR`` if it is not already present, then extracts it.
+    The zip contains a nested ``101_ObjectCategories.tar.gz`` archive
+    which is extracted in a second pass to produce the final category
+    directories.  Both archives are deleted after extraction to reclaim
+    disk space.
+
+    Args:
+        on_progress: Optional progress callback. Falls back to the
+            application-wide ``update_progress`` when ``None``.
+
+    Returns:
+        Path to the ``101_ObjectCategories/`` directory containing category
+        subfolders of JPEG images (e.g.
+        ``data/caltech-101/101_ObjectCategories``).
+    """
+    if on_progress is None:
+        on_progress = _default_progress()
+
+    zip_path = DATA_DIR / "caltech-101.zip"
+    extract_dir = DATA_DIR / "caltech-101"
+    DATA_DIR.mkdir(exist_ok=True)
+    IMAGE_DIR.mkdir(exist_ok=True, parents=True)
+
+    categories_dir = extract_dir / "101_ObjectCategories"
+    if not categories_dir.exists():
+        if not zip_path.exists():
+            on_progress("downloading", "Starting Caltech-101 download...", 0, 0)
+            download_file_with_progress(
+                CALTECH101_URL, zip_path, CALTECH101_DOWNLOAD_SIZE_MB * 1024 * 1024, on_progress
+            )
+
+        on_progress("downloading", "Extracting Caltech-101 zip...", 0, 0)
+        with zipfile.ZipFile(zip_path, "r") as zip_ref:
+            members = zip_ref.namelist()
+            total = len(members)
+            for i, member in enumerate(members, 1):
+                if i % 100 == 0 or i == total:
+                    on_progress(
+                        "downloading",
+                        f"Extracting Caltech-101 zip ({i}/{total})...",
+                        i,
+                        total,
+                    )
+                zip_ref.extract(member, DATA_DIR)
+
+        zip_path.unlink(missing_ok=True)
+
+        # The zip contains 101_ObjectCategories.tar.gz (a nested archive).
+        # Extract it to produce the actual category directories.
+        inner_tar = extract_dir / "101_ObjectCategories.tar.gz"
+        if inner_tar.exists() and not categories_dir.exists():
+            on_progress("downloading", "Extracting 101_ObjectCategories...", 0, 0)
+            with tarfile.open(inner_tar, "r:gz") as tar_ref:
+                members = tar_ref.getmembers()
+                total = len(members)
+                for i, member in enumerate(members):
+                    if i % 100 == 0 or i == total - 1:
+                        on_progress(
+                            "downloading",
+                            f"Extracting 101_ObjectCategories ({i + 1}/{total})...",
+                            i + 1,
+                            total,
+                        )
+                    tar_ref.extract(member, extract_dir, filter="data")
+            inner_tar.unlink(missing_ok=True)
+
+    return categories_dir
+
+
+def download_caltech256(on_progress: Optional[ProgressCallback] = None) -> Path:
+    """Download and extract the Caltech-256 image classification dataset.
+
+    Downloads ``256_ObjectCategories.tar`` from the configured
+    ``CALTECH256_URL`` into ``DATA_DIR`` if it is not already present, then
+    extracts it.  The archive is deleted after extraction to reclaim disk
+    space.
+
+    Args:
+        on_progress: Optional progress callback. Falls back to the
+            application-wide ``update_progress`` when ``None``.
+
+    Returns:
+        Path to the ``256_ObjectCategories/`` directory containing category
+        subfolders of JPEG images (e.g.
+        ``data/caltech-256/256_ObjectCategories``).
+    """
+    if on_progress is None:
+        on_progress = _default_progress()
+
+    IMAGE_DIR.mkdir(exist_ok=True, parents=True)
+
+    categories_dir = DATA_DIR / "caltech-256" / "256_ObjectCategories"
+    _download_and_extract(
+        url=CALTECH256_URL,
+        archive_name="256_ObjectCategories.tar",
+        extract_to=DATA_DIR / "caltech-256",
+        check_path=categories_dir,
+        download_size_mb=CALTECH256_DOWNLOAD_SIZE_MB,
+        dataset_name="Caltech-256",
+        on_progress=on_progress,
+    )
+    return categories_dir
+
+
+def download_oxford_flowers(on_progress: Optional[ProgressCallback] = None) -> Path:
+    """Download and extract the Oxford Flowers 102 dataset.
+
+    Downloads the image tarball and the labels MAT file from the Oxford VGG
+    website into ``DATA_DIR`` if they are not already present, then extracts
+    the images.  Archives are deleted after extraction to reclaim disk space.
+
+    The dataset contains 8189 images of 102 flower species.  Images are stored
+    in a single flat directory (``jpg/``) with numeric filenames; the class
+    label for each image is provided in a separate MATLAB ``.mat`` file.
+
+    Args:
+        on_progress: Optional progress callback.  Falls back to the
+            application-wide ``update_progress`` when ``None``.
+
+    Returns:
+        Path to the ``oxford_flowers/`` directory containing ``jpg/`` and
+        ``imagelabels.mat`` (e.g. ``data/oxford_flowers``).
+    """
+    if on_progress is None:
+        on_progress = _default_progress()
+
+    IMAGE_DIR.mkdir(exist_ok=True, parents=True)
+
+    extract_dir = DATA_DIR / "oxford_flowers"
+    _download_and_extract(
+        url=OXFORD_FLOWERS_URL,
+        archive_name="102flowers.tgz",
+        extract_to=extract_dir,
+        check_path=extract_dir / "jpg",
+        download_size_mb=OXFORD_FLOWERS_DOWNLOAD_SIZE_MB,
+        dataset_name="Oxford Flowers",
+        on_progress=on_progress,
+    )
+
+    # Download labels file if not present.
+    labels_path = extract_dir / "imagelabels.mat"
+    if not labels_path.exists():
+        DATA_DIR.mkdir(exist_ok=True)
+        on_progress("downloading", "Downloading Oxford Flowers labels...", 0, 0)
+        download_file_with_progress(OXFORD_FLOWERS_LABELS_URL, labels_path, 1024 * 1024, on_progress)
+
+    return extract_dir
+
+
+def download_food101(on_progress: Optional[ProgressCallback] = None) -> Path:
+    """Download and extract the Food-101 dataset.
+
+    Downloads ``food-101.tar.gz`` from ETH Zurich into ``DATA_DIR`` if it is
+    not already present, then extracts it.  The archive is deleted after
+    extraction to reclaim disk space.
+
+    The dataset contains 101 000 images across 101 food categories (1000
+    images per category), stored in category subdirectories.
+
+    Args:
+        on_progress: Optional progress callback.  Falls back to the
+            application-wide ``update_progress`` when ``None``.
+
+    Returns:
+        Path to the ``food-101/images/`` directory containing category
+        subdirectories with JPEG images (e.g. ``data/food-101/images``).
+    """
+    if on_progress is None:
+        on_progress = _default_progress()
+
+    IMAGE_DIR.mkdir(exist_ok=True, parents=True)
+
+    images_dir = DATA_DIR / "food-101" / "images"
+    _download_and_extract(
+        url=FOOD101_URL,
+        archive_name="food-101.tar.gz",
+        extract_to=DATA_DIR,
+        check_path=images_dir,
+        download_size_mb=FOOD101_DOWNLOAD_SIZE_MB,
+        dataset_name="Food-101",
+        on_progress=on_progress,
+    )
+    return images_dir
+
+
+def download_eurosat(on_progress: Optional[ProgressCallback] = None) -> Path:
+    """Download and extract the EuroSAT RGB dataset.
+
+    Downloads ``EuroSAT_RGB.zip`` from HuggingFace into ``DATA_DIR`` if it
+    is not already present, then extracts it.  The archive is deleted after
+    extraction to reclaim disk space.
+
+    The dataset contains 27 000 Sentinel-2 satellite image patches across
+    10 land-use classes, stored in class subdirectories.
+
+    Args:
+        on_progress: Optional progress callback.  Falls back to the
+            application-wide ``update_progress`` when ``None``.
+
+    Returns:
+        Path to the ``EuroSAT_RGB/`` directory containing class
+        subdirectories with JPEG images (e.g. ``data/EuroSAT_RGB``).
+    """
+    if on_progress is None:
+        on_progress = _default_progress()
+
+    IMAGE_DIR.mkdir(exist_ok=True, parents=True)
+
+    extract_dir = DATA_DIR / "EuroSAT_RGB"
+    _download_and_extract(
+        url=EUROSAT_URL,
+        archive_name="EuroSAT_RGB.zip",
+        extract_to=DATA_DIR,
+        check_path=extract_dir,
+        download_size_mb=EUROSAT_DOWNLOAD_SIZE_MB,
+        dataset_name="EuroSAT",
+        on_progress=on_progress,
+    )
+    return extract_dir
+
+
+def download_stanford_dogs(on_progress: Optional[ProgressCallback] = None) -> Path:
+    """Download and extract the Stanford Dogs dataset.
+
+    Downloads ``images.tar`` from the Stanford Vision Lab into ``DATA_DIR``
+    if it is not already present, then extracts it.  The archive is deleted
+    after extraction to reclaim disk space.
+
+    The dataset contains ~20 580 images across 120 dog breed classes.  The
+    archive extracts to an ``Images/`` directory with breed subdirectories
+    named like ``n02085620-Chihuahua``.
+
+    Args:
+        on_progress: Optional progress callback.  Falls back to the
+            application-wide ``update_progress`` when ``None``.
+
+    Returns:
+        Path to the ``stanford_dogs/Images/`` directory containing breed
+        subdirectories with JPEG images (e.g.
+        ``data/stanford_dogs/Images``).
+    """
+    if on_progress is None:
+        on_progress = _default_progress()
+
+    IMAGE_DIR.mkdir(exist_ok=True, parents=True)
+
+    images_dir = DATA_DIR / "stanford_dogs" / "Images"
+    _download_and_extract(
+        url=STANFORD_DOGS_URL,
+        archive_name="stanford_dogs_images.tar",
+        extract_to=DATA_DIR / "stanford_dogs",
+        check_path=images_dir,
+        download_size_mb=STANFORD_DOGS_DOWNLOAD_SIZE_MB,
+        dataset_name="Stanford Dogs",
+        on_progress=on_progress,
+    )
+    return images_dir
